put manual under bootstrap
This is a slightly manually tweaked version from the markdown export in the
main repo. Still need to automate the TOC generation better.

Also had to adjust h1 and h2 in jflex.css to make them work as link targets
(#content padding seems to only work for the #content link itself).